Revised Launch descriptions
I've revised the descriptions for both EHR and Standalone launches, adding new sequence diagrams to each.  

As part of this, I moved some of the detail that previously had been in the overview down into the descriptions.
Some of the text relating to conformance statements seemed inconsistent and didn't align with the sequence diagrams.  I've tried to clean that up.