Removed detail and sequence diagram from overview
Previously I pulled the detail re EHR launch and stand-alone launch out of the overview and moved them into the launch descriptions.  But I neglected to delete the moved content from the overview.  This change does that.

Also, we're replacing the single, complex sequence diagrams with 4 separate sequences.  So I have deleted it from the overview.