Typo corrections and comments
1)  Corrected some typos.

2)  Deleted "(but note that requiring client authentication via mutual TLS may limit
interoperability of the client across server environments)."  I have several concerns about this phrase.  First, it tends to imply that a policy that requires a client to authenticate itself is bad policy -- which isn't necessarily the case, and who are we to tell organizations what their policy should be?  Second, maximizing the interoperability of a client is not the responsibility of an authorization server, or the organization holding the resource.  Third (related to 2), just because one authorization server requires a client to authenticate itself doesn't mean that every authorization server needs to do so.  

3) Deleted the question re 2.1.2, with my response --  I don’t think we should limit the applicability of our best practices to use-case 1-4 (patient/provider client access).  So I don’t see this as “hinting” at our cross-organizational approach.    As for detail, I tried not to duplicate the content in the OAuth spec on this model.  In what areas do you think more detail would be appropriate?

4) In 2.4.1, changed "To accomplish this" to "To fulfill this requirement" -- not clear what "this" referred to.